fix types

+++ src/igor/Core/Errorrate.h
@@ -172,7 +172,7 @@ public:
     virtual int get_number_non_zero_likelihood_seqs() const = 0;
     virtual std::queue<int> generate_errors(std::string &, std::mt19937_64 &) const = 0;
     void set_viterbi_run(bool viterbi_like) { viterbi_run = viterbi_like; }
-    int debug_number_scenarios;
+    size_t debug_number_scenarios;
 
 protected:
     bool updated;
